+++ addon/data-export.js
@@ -102,7 +102,6 @@ class Model {
     this.autocompleteProgress = {};
     this.exportProgress = {};
     this.queryName = "";
-    this.clientId = localStorage.getItem(sfHost + "_clientId") ? localStorage.getItem(sfHost + "_clientId") : "";
     this.queryTemplates = localStorage.getItem("queryTemplates") ? this.queryTemplates = localStorage.getItem("queryTemplates").split("//") : [
       "SELECT Id FROM ",
       "SELECT Id FROM WHERE",
@@ -145,9 +144,6 @@ class Model {
   }
   setQueryName(value) {
     this.queryName = value;
-  }
-  setClientId(value) {
-    this.clientId = value;
   }
   setQueryInput(queryInput) {
     this.queryInput = queryInput;
@@ -250,9 +246,6 @@ class Model {
   }
   addToHistory() {
     this.savedHistory.add({query: this.getQueryToSave(), useToolingApi: this.queryTooling});
-  }
-  saveClientId() {
-    localStorage.setItem(this.sfHost + "_clientId", this.clientId);
   }
   removeFromHistory() {
     this.savedHistory.remove({query: this.getQueryToSave(), useToolingApi: this.queryTooling});
@@ -1035,7 +1028,6 @@ class App extends React.Component {
     this.onClearHistory = this.onClearHistory.bind(this);
     this.onSelectSavedEntry = this.onSelectSavedEntry.bind(this);
     this.onAddToHistory = this.onAddToHistory.bind(this);
-    this.onSaveClientId = this.onSaveClientId.bind(this);
     this.onRemoveFromHistory = this.onRemoveFromHistory.bind(this);
     this.onClearSavedHistory = this.onClearSavedHistory.bind(this);
     this.onToggleHelp = this.onToggleHelp.bind(this);
@@ -1050,7 +1042,6 @@ class App extends React.Component {
     this.onDeleteRecords = this.onDeleteRecords.bind(this);
     this.onResultsFilterInput = this.onResultsFilterInput.bind(this);
     this.onSetQueryName = this.onSetQueryName.bind(this);
-    this.onSetClientId = this.onSetClientId.bind(this);
     this.onStopExport = this.onStopExport.bind(this);
   }
   onQueryAllChange(e) {
@@ -1095,12 +1086,6 @@ class App extends React.Component {
     e.preventDefault();
     let {model} = this.props;
     model.addToHistory();
-    model.didUpdate();
-  }
-  onSaveClientId(e) {
-    e.preventDefault();
-    let {model} = this.props;
-    model.saveClientId();
     model.didUpdate();
   }
   onRemoveFromHistory(e) {
@@ -1189,11 +1174,6 @@ class App extends React.Component {
   onSetQueryName(e) {
     let {model} = this.props;
     model.setQueryName(e.target.value);
-    model.didUpdate();
-  }
-  onSetClientId(e) {
-    let {model} = this.props;
-    model.setClientId(e.target.value);
     model.didUpdate();
   }
   onStopExport() {
@@ -1320,9 +1300,7 @@ class App extends React.Component {
               ),
               h("input", {placeholder: "Query Label", type: "save", value: model.queryName, onInput: this.onSetQueryName}),
               h("button", {onClick: this.onAddToHistory, title: "Add query to saved history"}, "Save Query"),
-              h("button", {className: model.expandSavedOptions ? "toggle contract" : "toggle expand", title: "Show More Options", onClick: this.onToggleSavedOptions}, h("div", {className: "button-toggle-icon"})),
-              h("input", {placeholder: "Consumer Key", type: "default", value: model.clientId, onInput: this.onSetClientId}),
-              h("button", {onClick: this.onSaveClientId, title: "Save Consumer Key"}, "Save"),
+              h("button", {className: model.expandSavedOptions ? "toggle contract" : "toggle expand", title: "Show More Options", onClick: this.onToggleSavedOptions}, h("div", {className: "button-toggle-icon"}))
             ),
           ),
           h("div", {className: "query-options"},